Code Structure Remastered
New - Defs packages
New - Packages structured project
New - SocialFish was replaced by HiddenEye
