+++ Defs/Configurations.py
@@ -1,11 +1,15 @@
 #Primitive config works
 import configparser
+import os
 
 def createConfig(path = "Settings.ini"):
     config = configparser.ConfigParser()
     config.add_section("Settings")
+    config.add_section("Defaults")
     config.set("Settings", "Language", "en")
-
+    config.set("Settings", "DidBackground", "True")
+    config.set("Defaults", "webPage", "Facebook")
+    config.set("Defaults", "additionalOption", "1")
     with open(path, 'w') as configFile:
         config.write(configFile)
 
@@ -13,3 +17,7 @@ def readConfig(path = "Settings.ini"):
     config = configparser.ConfigParser()
     config.read(path)
     return config
+
+def ifSettingsNotExists():
+    if not os.path.exists("Settings.ini"):
+        createConfig()
